ISSUE6 근태관리 (Attendance Management) - 휴가 관리 - 출근 시간 체크 - 패널티 기능 - 스케줄 관리 기능 issue7에서 이어서 개발

+++ src/main/java/com/commit/collaboration_board_server/model/MonthlyWorkDate.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class MonthlyWorkDate {
     private Long id;
     private String userId;
-    private int totalWorkTime;
+    private Double totalWorkTime;
 }
